renamed AirborneVelocityMsg to match convention
