+++ src/main/java/de/serosystems/lib1090/msgs/adsb/VelocityOverGroundMsg.java
@@ -28,7 +28,7 @@ import java.io.Serializable;
  * Decoder for ADS-B velocity messages
  * @author Matthias Schäfer (schaefer@sero-systems.de)
  */
-public class VelocityOverGroundMsg extends ExtendedSquitter implements Serializable, AirborneVelocityMessage {
+public class VelocityOverGroundMsg extends ExtendedSquitter implements Serializable, AirborneVelocityMsg {
 
 	private static final long serialVersionUID = 1774082354457574555L;
