+++ src/test/java/de/serosystems/lib1090/msgs/adsb/AirspeedHeadingMsgTest.java
@@ -55,7 +55,7 @@ class AirspeedHeadingMsgTest {
 	@Test
 	public void testImplementsAirborneVelocityMessage() throws Exception {
 		AirspeedHeadingMsg msg = new AirspeedHeadingMsg("8DA05F219B06B6AF189400CBC33F");
-		assertInstanceOf(AirborneVelocityMessage.class, msg);
+		assertInstanceOf(AirborneVelocityMsg.class, msg);
 	}
 
 	@Test
